# check-yo-self-api/sdk-swagger.json
{
  "x-generator": "NSwag v13.0.4.0 (NJsonSchema v10.0.21.0 (Newtonsoft.Json v11.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "check-yo-self-api",
    "version": "1.0.0"
  },
  "paths": {
    "/api/Employees": {
      "get": {
        "tags": [
          "Employees"
        ],
        "operationId": "Employees_GetAll",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Employee"
              }
            }
          },
          "500": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Employees"
        ],
        "operationId": "Employees_Post",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "employee",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Employee"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          },
          "500": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      }
    },
    "/api/Employees/{employeeId}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "operationId": "Employees_GetById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "employeeId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          },
          "500": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Employees"
        ],
        "operationId": "Employees_Delete",
        "parameters": [
          {
            "type": "integer",
            "name": "employeeId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Employees"
        ],
        "operationId": "Employees_Update",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "employeeId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "employee",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Employee"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/api/Employees/SalaryGreaterEqualTo/{salary}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "operationId": "Employees_GetBySalary",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "number",
            "name": "salary",
            "in": "path",
            "required": true,
            "format": "decimal",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Employee"
              }
            }
          },
          "500": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      }
    },
    "/api/Employees/GetByLastName/{lastName}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "operationId": "Employees_GetByLastName",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "lastName",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Employee"
              }
            }
          },
          "500": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      }
    },
    "/api/Employees/GetByFirstName/{firstName}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "operationId": "Employees_GetByFirstName",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "firstName",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Employee"
              }
            }
          },
          "500": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      }
    },
    "/api/Employees/GetByFullName/{firstName}/{lastName}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "operationId": "Employees_GetByFullName",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "firstName",
            "in": "path",
            "required": true,
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "lastName",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Employee"
              }
            }
          },
          "500": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/Employees/ReindexAllEmployees": {
      "post": {
        "tags": [
          "Employees"
        ],
        "operationId": "Employees_ReindexAllEmployees",
        "responses": {
          "200": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/api/GettingStarted": {
      "get": {
        "tags": [
          "GettingStarted"
        ],
        "operationId": "GettingStarted_Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GettingStarted"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Employee": {
      "type": "object",
      "required": [
        "employeeId",
        "lastName",
        "firstName",
        "salary",
        "firstPaycheckDate"
      ],
      "properties": {
        "employeeId": {
          "type": "integer",
          "format": "int32"
        },
        "lastName": {
          "type": "string",
          "maxLength": 1024,
          "minLength": 0
        },
        "firstName": {
          "type": "string",
          "maxLength": 1024,
          "minLength": 0
        },
        "salary": {
          "type": "number",
          "format": "decimal"
        },
        "firstPaycheckDate": {
          "type": "string",
          "format": "date-time",
          "minLength": 1
        }
      }
    },
    "GettingStarted": {
      "type": "object",
      "properties": {
        "applicationName": {
          "type": "string"
        }
      }
    }
  }
}